m feat: added Replayer support for arrival trace j feat: added JMTOptions, MVAOptions, etc j feat: added Replayer support for arrival trace j feat: several LN bug fixes p refactor: updated getting started examples

+++ java/src/main/java/jline/solvers/Solver.java
@@ -66,7 +66,7 @@ public abstract class Solver {
   }
 
   // Assign a new seed to the random number generator
-  public void resetRandomGeneratorSeed(int seed) {
+  public void resetRandomGeneratorSeed(long seed) {
     this.random = new Random(seed);
   }
